集成quartz集群，任务分配责任链

+++ springbootstudydomain/src/main/java/com/baili/dao/MemberMapper.java
@@ -7,7 +7,7 @@ import org.apache.ibatis.annotations.Param;
 import org.apache.ibatis.session.RowBounds;
 
 public interface MemberMapper {
-    int countByExample(MemberCriteria example);
+    long countByExample(MemberCriteria example);
 
     int deleteByExample(MemberCriteria example);
